Chatbot/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.54.0
faiss-cpu==1.8.0
numpy==1.26.4
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
tiktoken==0.7.0
requests==2.32.3
aiofiles==24.1.0
